Add an email template for the weekly notifier.
It's super basic but includes a pie chart for a nice visualization of
how the time is split. We are using Google's static chart API which has
a big warning saying that it is deprecated but also saying they have no
plans to actually kill it. So it will hopefully be good for a while.