Use Hourglass's data to determine when to send timer reminders
Rather than making an external call to Harvest, we can rely on the data
we have being up to date enough to determine whether a user has started
a timer yet today.

This also removes the ability for this particular interactor to be
callable with a custom date because it doesn't make as much sense in
this context as it does for an interactor like FetchDailyHours.

The change also fixes a bug where a user's timer_reminder_sent_on column
was updated even if no reminder was necessary.