Remove the unassigned entry and only show hours for assigned users
While the system does pull in all time entries for the given Harvest
project, the "unassigned" setup didn't really make much sense in
practice. Instead, we will only report on the hours for users who are
assigned, and will compare the team hours only to the combined hours of
assigned users.